koding/config: move to koding/kites/config
