+++ go/src/koding/kites/config/config.json
